+++ src/riccati/newton.jl
@@ -1,38 +1,33 @@
 # This file is a part of DifferentialRiccatiEquations. License is MIT: https://spdx.org/licenses/MIT.html
 
-# Keep defaults in sync with docstring of NewtonADI!
 function CommonSolve.solve(
     prob::GAREProblem{TG,TQ},
-    ::NewtonADI;
-    reltol = size(prob.A, 1) * eps(),
-    maxiters = 5,
+    alg::Newton;
     observer = nothing,
-    adi_initprev::Bool = false,
-    adi_kwargs = NamedTuple(),
-    inexact::Bool = true,
-    inexact_hybrid::Bool = true,
-    inexact_forcing = quadratic_forcing,
-    linesearch::Bool = true,
 ) where {TG,TQ}
     TG <: LDLᵀ{<:AbstractMatrix,UniformScaling{Bool}} || error("TG=$TG not yet implemented")
     TQ <: LDLᵀ{<:AbstractMatrix,UniformScaling{Bool}} || error("TQ=$TQ not yet implemented")
 
-    @timeit_debug "callbacks" observe_gare_start!(observer, prob, NewtonADI())
+    @timeit_debug "callbacks" observe_gare_start!(observer, prob, alg)
     TL = TD = Matrix{Float64}
 
     @unpack E, A, Q = prob
     B, _ = prob.G
     Cᵀ, _ = Q
+    res = Q
+    res_norm = norm(res)
+    reltol = @something(alg.reltol, size(A, 1) * eps())
+    abstol = @something(alg.abstol, reltol * res_norm)
 
     n = size(A, 2)
     X = LDLᵀ{TL,TD}(zeros(n, 0), zeros(0, 0)) # this is ugly
 
-    res = Q
-    res_norm = norm(res)
-    abstol = reltol * res_norm
-
     i = 0
     local X_prev
+    @unpack maxiters, inner_alg = alg
+    inner_reltol = unpack(inner_alg, Val(:reltol))
+    inner_reltol = @something(inner_reltol, reltol / 10)
+    @unpack inexact, inexact_hybrid, inexact_forcing, linesearch = alg
     while true
         # Compute residual
         L, D = X
@@ -91,7 +86,7 @@ function CommonSolve.solve(
         res_norm <= abstol && break
         if i >= maxiters
             @timeit_debug "callbacks" observe_gare_failed!(observer)
-            @warn "NewtonADI did not converge" residual=res_norm abstol maxiters
+            @warn "Newton method did not converge" residual=res_norm abstol maxiters
             break
         end
         i += 1
@@ -107,39 +102,36 @@ function CommonSolve.solve(
         S::TD = _dcat(TD, I(q), I(m))
         RHS = LDLᵀ(G, S)
 
-        # ADI setup
+        # Lyapunov setup
         lyap = GALEProblem(E, F, RHS)
-        initial_guess = adi_initprev ? X : nothing
-        adi_reltol = get(adi_kwargs, :reltol, reltol / 10)
         if inexact
             η = inexact_forcing(i, res_norm)
-            adi_abstol = η * res_norm
+            inner_abstol = η * res_norm
             if inexact_hybrid
                 # If the classical/"exact" tolerance is less strict than
                 # the one of the Inexact Newton, use that tolerance instead.
-                classical_abstol = adi_reltol * norm(lyap.C)
-                switch_back = classical_abstol > adi_abstol
+                classical_abstol = inner_reltol * norm(lyap.C)
+                switch_back = classical_abstol > inner_abstol
                 @timeit_debug "callbacks" observe_gare_metadata!(observer, "inexact", !switch_back)
                 if switch_back
-                    @debug "Switching from inexact to classical Newton method" i inexact_abstol=adi_abstol classical_abstol
-                    adi_abstol = classical_abstol
+                    @debug "Switching from inexact to classical Newton method" i inexact_abstol=inner_abstol classical_abstol
+                    inner_abstol = classical_abstol
                 end
             else
                 @timeit_debug "callbacks" observe_gare_metadata!(observer, "inexact", true)
             end
         else
-            adi_abstol = adi_reltol * norm(lyap.C)
+            inner_abstol = inner_reltol * norm(lyap.C)
         end
 
         # Newton step:
         X_prev = X
-        X = @timeit_debug "ADI" solve(
-            lyap, ADI();
-            maxiters=100,
+        X = @timeit_debug "$(typeof(inner_alg))" solve(
+            lyap, inner_alg;
+            abstol=inner_abstol,
+            initial_guess=X_prev,
             observer,
-            initial_guess,
-            abstol=adi_abstol,
-            adi_kwargs...)
+        )
     end
 
     @timeit_debug "callbacks" observe_gare_done!(observer, i, X, res, res_norm)
@@ -151,7 +143,7 @@ end
 
 Exemplary forcing term to obtain superlinear convergence in the inexact Newton method.
 `i::Int` refers to the current Newton step.
-See [`NewtonADI`](@ref).
+See [`Newton`](@ref).
 """
 superlinear_forcing(i, _) = 1 / (i^3 + 1)
 
@@ -160,6 +152,6 @@ superlinear_forcing(i, _) = 1 / (i^3 + 1)
 
 Exemplary forcing term to obtain quadratic convergence in the inexact Newton method.
 `residual_norm::Float64` refers to the norm of the previous Newton residual.
-See [`NewtonADI`](@ref).
+See [`Newton`](@ref).
 """
 quadratic_forcing(_, residual_norm) = min(0.1, 0.9 * residual_norm)